pyproject.toml
[project]
name = "azure-ai-agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=2.1.0",
    "azure-cosmos>=4.15.0",
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.7",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
    "plotly>=6.7.0",
    "pypdf>=6.10.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "sec-edgar-downloader>=5.1.0",
    "streamlit>=1.57.0",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "ruff>=0.15.12",
]
